pyproject.toml
[tool.poetry]
name = "dionysus"
version = "0.1.0"
description = ""
authors = ["shinkeonkim <singun11@kookmin.ac.kr>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.5"
djangorestframework = "^3.15.2"
django-filtering = "^0.0.5"
markdown = "^3.7"
drf-spectacular = "^0.28.0"
django-safedelete = "^1.4.0"
drf-extensions = "^0.7.1"
django-filter = "^24.3"
django-guid = "^3.5.0"
drf-api-logger = "^1.1.16"
django-lifecycle = "^1.2.4"
drf-writable-nested = "^0.7.1"
django-admin-autocomplete-filter = "^0.7.1"
django-unfold = "^0.46.0"
django-import-export = "^4.3.4"
django-guardian = "^2.4.0"
django-simple-history = "^3.7.0"
django-markdownx = "^4.0.7"
django-environ = "^0.12.0"
psycopg2-binary = "^2.9.10"
dj-rest-auth = "^7.0.1"
django-allauth = "^65.3.1"
djangorestframework-simplejwt = "^5.4.0"
requests = "^2.32.3"
cryptography = "^44.0.0"
drf-nested-routers = "^0.94.1"
nplusone = "^1.0.0"
django-silk = "^5.3.2"

[tool.poetry.dev-dependencies]
pre-commit = "^4.0.1"
black = "^24.10.0"
flake8 = "^7.1.1"
isort = "^5.9.3"
mypy = "^1.8.0"
autoflake = "^2.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
